finish moving metrics
