+++ dataplane-webserver/src/metrics/types.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Clone)]
 pub struct RangeQuery {
     pub query: String,
     pub start: f64,
